looks like a TypeInType bug
commenting out `Branch''`
makes the bug go away

BUT adding a new axiom to an existing proposition shouldn't
invalidate the proposition!

`Branch'` should compile regardless of the existence of
`Branch''` in `HTree''`.

Conversely I get error-free compilation when Commenting out `Branch'`
(not sure if this is a sign of symmetry)